pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cscn"
version = "0.1.0"
description = "Generic CSCN workflows for scRNA-seq datasets."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "Allen Y. G. Yang"}
]
dependencies = [
  "numpy>=1.24,<3.0",
  "pandas>=2.0,<3.0",
  "scipy>=1.10,<2.0",
  "scikit-learn>=1.3,<2.0",
  "networkx>=3.0,<4.0",
  "pgmpy>=0.1.26,<=1.0",
  "PyYAML>=6.0,<7.0",
  "anndata>=0.10,<1.0",
]

[project.optional-dependencies]
viewer = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.32,<1.0",
]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
cscn = "cscn.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
